Reconstruct Nawatobi level menu and script transitions

+++ include/game/debug_console.h
@@ -13,7 +13,7 @@ typedef struct GameConsole {
 
 void GameConsole_Write(GameConsole *console, const char *text);
 void GameConsole_Printf(GameConsole *console, const char *format, ...);
-void GameConsole_SetCursor(GameConsole *console, u8 x, u8 y);
+void GameConsole_SetCursor(GameConsole *console, int x, int y);
 void GameConsole_Upload(GameConsole *console);
 void GameConsole_Clear(GameConsole *console);
 GameConsole *GameConsole_Init(GameConsole *console, int sub_screen, int background,
